+++ src/twg2/io/files/Bytes.java
@@ -1,6 +1,6 @@
 package twg2.io.files;
 
-/** A utility class for converting primitive types to bytes.
+/** A utility class for converting primitive types to/from bytes in big-endian order.
  * For example converting an integer to 4 bytes and storing those 4 bytes
  * at a specific location in a byte array.
  * @author TeamworkGuy2
@@ -10,32 +10,33 @@ public final class Bytes {
 
 	private Bytes() { throw new AssertionError("cannot instantiate Bytes"); }
 
-	/** Write a double to the specified byte array as 8 bytes
+
+	/** Write a double to the specified byte array as 8 bytes (big-endian order)
 	 * @param value the double to write
-	 * @param b the byte array to write the long to
+	 * @param b the byte array to write the 'value' to
 	 * @param offset the offset into the array at which to write the 8 bytes
 	 */
-	public static final void writeDouble(double value, byte[] b, int offset) {
+	public static void writeDouble(double value, byte[] b, int offset) {
 		writeLong(Double.doubleToRawLongBits(value), b, offset);
 	}
 
 
-	/** Write a float to the specified byte array as 4 bytes
+	/** Write a float to the specified byte array as 4 bytes (big-endian order)
 	 * @param value the float to write
-	 * @param b the byte array to write the float to
+	 * @param b the byte array to write the 'value' to
 	 * @param offset the offset into the array at which to write the 4 bytes
 	 */
-	public static final void writeFloat(float value, byte[] b, int offset) {
+	public static void writeFloat(float value, byte[] b, int offset) {
 		writeInt(Float.floatToRawIntBits(value), b, offset);
 	}
 
 
-	/** Write a long to the specified byte array as 8 bytes
+	/** Write a long to the specified byte array as 8 bytes (big-endian order)
 	 * @param value the long to write
-	 * @param b the byte array to write the long to
+	 * @param b the byte array to write the 'value' to
 	 * @param offset the offset into the array at which to write the 8 bytes
 	 */
-	public static final void writeLong(long value, byte[] b, int offset) {
+	public static void writeLong(long value, byte[] b, int offset) {
 		b[offset] = (byte)(value >>> 56);
 		b[offset+1] = (byte)(value >>> 48);
 		b[offset+2] = (byte)(value >>> 40);
@@ -47,12 +48,12 @@ public final class Bytes {
 	}
 
 
-	/** Write an int to the specified byte array as 4 bytes
+	/** Write an int to the specified byte array as 4 bytes (big-endian order)
 	 * @param value the integer to write
-	 * @param b the byte array to write the integer to
+	 * @param b the byte array to write the 'value' to
 	 * @param offset the offset into the array at which to write the 4 bytes
 	 */
-	public static final void writeInt(int value, byte[] b, int offset) {
+	public static void writeInt(int value, byte[] b, int offset) {
 		b[offset] = (byte)(value >>> 24);
 		b[offset+1] = (byte)(value >>> 16);
 		b[offset+2] = (byte)(value >>> 8);
@@ -60,39 +61,39 @@ public final class Bytes {
 	}
 
 
-	/** Write a short to the specified byte array as 2 bytes
+	/** Write a short to the specified byte array as 2 bytes (big-endian order)
 	 * @param value the short to write
-	 * @param b the byte array to write the short to
+	 * @param b the byte array to write the 'value' to
 	 * @param offset the offset into the array at which to write the 2 bytes
 	 */
-	public static final void writeShort(short value, byte[] b, int offset) {
+	public static void writeShort(short value, byte[] b, int offset) {
 		b[offset] = (byte)(value >>> 8);
 		b[offset+1] = (byte)(value );
 	}
 
 
-	/** Write a boolean to the specified byte array as 1 byte (1=true, 0=false)
+	/** Write a boolean to the specified byte array as 1 byte (1=true, 0=false) (big-endian order)
 	 * @param value the boolean to write
-	 * @param b the byte array to write the short to
+	 * @param b the byte array to write the 'value' to
 	 * @param offset the offset into the array at which to write the byte
 	 */
-	public static final void writeBoolean(boolean value, byte[] b, int offset) {
+	public static void writeBoolean(boolean value, byte[] b, int offset) {
 		b[offset] = (byte)(value ? 1 : 0);
 	}
 
 
-	/** Read a double value from the specified location in the specified array
+	/** Read a double value from the specified location in the specified array (assumes big-endian order)
 	 * @param b the array to read the double from
 	 * @param offset the offset into the array at which to read the 8 bytes
 	 * @return eight bytes read from the indices {@code [offset, offset+3]} and
 	 * converted to a double
 	 */
-	public static final double readDouble(byte[] b, int offset) {
+	public static double readDouble(byte[] b, int offset) {
 		return Double.longBitsToDouble(readLong(b, offset));
 	}
 
 
-	/** Read a float value from the specified location in the specified array
+	/** Read a float value from the specified location in the specified array (assumes big-endian order)
 	 * @param b the array to read the float from
 	 * @param offset the offset into the array at which to read the 4 bytes
 	 * @return four bytes read from the indices {@code [offset, offset+3]} and
@@ -103,7 +104,7 @@ public final class Bytes {
 	}
 
 
-	/** Read a long value from the specified location in the specified array
+	/** Read a long value from the specified location in the specified array (assumes big-endian order)
 	 * @param b the array to read the long from
 	 * @param offset the offset into the array at which to read the 8 bytes
 	 * @return eight bytes read from the indices {@code [offset, offset+3]} and converted to
@@ -124,7 +125,7 @@ public final class Bytes {
 	}
 
 
-	/** Read an integer value from the specified location in the specified array
+	/** Read an integer value from the specified location in the specified array (assumes big-endian order)
 	 * @param b the array to read the integer from
 	 * @param offset the offset into the array at which to read the 4 bytes
 	 * @return four bytes read from the indices {@code [offset, offset+3]} and converted to
@@ -135,7 +136,7 @@ public final class Bytes {
 	}
 
 
-	/** Read a short value from the specified location in the specified array
+	/** Read a short value from the specified location in the specified array (assumes big-endian order)
 	 * @param b the array to read the short from
 	 * @param offset the offset into the array at which to read the 2 bytes
 	 * @return two bytes read from indices {@code offset} and {@code offset+1} and converted to
@@ -146,7 +147,7 @@ public final class Bytes {
 	}
 
 
-	/** Read a boolean value from the specified location in the specified byte array (1=true, 0=false)
+	/** Read a boolean value from the specified location in the specified byte array (1=true, 0=false) (assumes big-endian order)
 	 * @param b the array to read the boolean from
 	 * @param offset the offset into the array at which to read the byte
 	 * @return two bytes read from indices {@code offset} and {@code offset+1} and converted to